refactor: move parsers into a dedicated file
