+++ applications_user/lora/tests/unit_tests/parsers_test.c
@@ -1,6 +1,7 @@
-#include "../../lora_app.h"
-
 #include <furi.h>
+#include <stdbool.h>
+
+#include "../../lora_parsers.h"
 
 /**
  * Test case for parsing a FPENDING message.
